Moved KeySwap and Debugger Allowance to EnginePatches
