+++ Game/Common.h
@@ -1,0 +1,2 @@
+#pragma once
+#define LOG(x) MessageBox(NULL, x, "RGSSAD Container", MB_ICONERROR)
