+++ Game/EnginePatches.h
@@ -2,4 +2,6 @@
 
 namespace Patches {
 void SetupDetours(const char* library_path);
+void SwapRgssadEncryption(const char* library_path);
+void PatchDebugPresent();
 } // namespace Patches
