+++ Game/Game.vcxproj
@@ -152,11 +152,10 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="AriMath.h" />
+    <ClInclude Include="Common.h" />
     <ClInclude Include="EnginePatches.h" />
     <ClInclude Include="ld32.h" />
     <ClInclude Include="MemoryUtil.h" />
-    <ClInclude Include="resource.h" />
-    <ClInclude Include="resource1.h" />
     <ClInclude Include="RubyCommon.h" />
     <ClInclude Include="SigScanner.h" />
   </ItemGroup>
